refactor: derive composer state from chat events
